backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.1
groq==0.4.2
websockets==12.0
python-multipart==0.0.9
